Fix for postgres startup wait
The official docker postgres container starts up twice and so waiting on
a single "ready to accept connections" string in the logs isn't
enough to know that the system is *actually* able to accept connections.
We now first wait for the postgres init process to complete before
watching for the string.